requirements.txt
requests
yfinance
beautifulsoup4
pandas
plotly
finance-datareader
yahooquery
python-telegram-bot
python-dotenv
